+++ src/components/AllMonsters.js
@@ -1,44 +1,9 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import styled from 'styled-components/macro'
-import saveLocally from './lib/saveLocally'
-import OneMonsterItem from './OneMonsterItem'
 import LoadingAnimation from './LoadingAnimation'
+import OneMonsterItem from './OneMonsterItem'
 
-export default function AllMonsters() {
-  const [monsterDetails, setMonsterDetails] = useState([])
-
-  const [loading, setLoading] = useState(false)
-
-  useEffect(() => {
-    const locallyLoadedMonsterNames = JSON.parse(
-      localStorage.getItem('monsternames')
-    )
-
-    if (!locallyLoadedMonsterNames || locallyLoadedMonsterNames.length === 0) {
-      setLoading(true)
-      fetch(`https://www.dnd5eapi.co/api/monsters`)
-        .then((res) => res.json())
-        .then((data) => {
-          saveLocally('monsternames', data.results)
-          data.results.forEach((monster) => {
-            fetch(`https://www.dnd5eapi.co${monster.url}`)
-              .then((res) => res.json())
-              .then((data) => {
-                setMonsterDetails((monsterDetails) => [...monsterDetails, data])
-              })
-              .then(setLoading(false))
-          })
-        })
-        .catch((err) => console.log(err))
-    }
-  }, [])
-
-  if (monsterDetails.length !== 0) {
-    saveLocally('monsterdetails', monsterDetails)
-  } else if (localStorage.getItem('monsterdetails')) {
-    setMonsterDetails(JSON.parse(localStorage.getItem('monsterdetails')))
-  }
-
+export default function AllMonsters({ loading, monsterDetails }) {
   if (loading) {
     return <LoadingAnimation />
   } else {
