+++ src/index.js
@@ -1,13 +1,16 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import { BrowserRouter as Router } from 'react-router-dom'
 import GlobalStyles from './GlobalStyles'
 import App from './App'
 import * as serviceWorker from './serviceWorker'
 
 ReactDOM.render(
   <React.StrictMode>
-    <GlobalStyles />
-    <App />
+    <Router>
+      <GlobalStyles />
+      <App />
+    </Router>
   </React.StrictMode>,
   document.getElementById('root')
 )
